+++ CHANGELOG.md
@@ -12,6 +12,7 @@ contents to a new release section (e.g. _[0.x]_).
 ### Added
 
 - Added a checkbox to Poker to toggle between solver and JSON math formats (PLUT-691)
+- Added support for multivariate linear inequalities, such as `3a + 2b > 9` (PLUT-693)
 
 ### Fixed